perf(profile): long-context attribution report + tooling (#18)
* fix(attn): K/V dtype bug in three Metal attention kernels + parity tests

flash_attention.metal, attention_gqa.metal, and attention_splitk.metal
(compute kernel) bound K/V caches as `const device float*`, while the
caches are stored as f16 on the Rust side (encode_cast_f32_to_f16_offsets,
see the `// f16 bytes` comment in metal_lfm2.rs:2445). Reading f16 as f32
reinterprets two adjacent halves as one f32 — every KV load in these
kernels was garbage.

Impact:
- flash_attention auto-activates at seq_len > 4096 (metal_lfm2.rs:1626),
  so any context window past 4096 produced corrupt logits and silently
  unusable generation. This is the "14% slower than classic, needs
  rewrite" item in the memory note — not just slower, actually broken.
- attention_gqa is opt-in via WICK_ATTN=gqa (latent, but silently wrong
  whenever opted in).
- attention_splitk is opt-in via WICK_ATTN=splitk (same).
- Classic attention.metal and attention_prefill.metal bound correctly as
  half* — only these two were safe.

Fix: change K/V bindings to `half*` and cast to float on load (3 edits
per file).

Tests:
- wick/tests/metal_shaders_parity.rs: 3 new synthetic-input tests
  (test_classic_vs_{flash,gqa,splitk}_attention_synthetic) that run in
  CI — no model file needed. Construct Q/K/V buffers directly, run each
  kernel variant, compare outputs against classic (always-correct
  half* binding) with 1e-3 tolerance.
- wick/tests/attention_metal_parity.rs (new): 3 ignored end-to-end
  tests that reload the model with WICK_FLASH=1 / WICK_ATTN=gqa /
  WICK_ATTN=splitk and assert the greedy decode matches classic on a
  real 450M-Q4_0 model. Env-var mutations guarded by an RAII
  EnvVarGuard backed by a process-wide OnceLock<Mutex<()>>, restoring
  the prior value on Drop (panic-safe) and serializing concurrent
  tests.

All 16 metal_shaders_parity tests pass (13 pre-existing + 3 new). All
3 end-to-end parity tests pass post-fix (pre-fix they fail with
classic=6 coherent tokens vs variant=12 random tokens).

* perf(profile): long-context attribution report + tooling

Profile pass to attribute wick's long-context regression vs llama.cpp
(benchmarks/deltas_table.md: prefill 0.22x at p=4096, decode 0.28-0.54x
at ctx=4096).

Findings (full report: benchmarks/profile_longctx.md):
- Decode scaling on LFM2.5-VL-450M-Q4_0: attn_kernel goes from 7.7%
  of per-step time at ctx=128 (1.60 ms) to 54% at ctx=4096 (23.65 ms).
  All other phases are flat (1.0-1.2x scaling). Caveat: ctx<=2048 uses
  classic `attention.metal`, ctx=4096 crosses seq_len > 4096 and uses
  flash — part of the 14.8x jump is the kernel switch.
- Prefill batched attribution via forward_prefill_profiled shows the
  same signature — attn_kernel is the only phase whose per-token cost
  grows with n (8.8x for 32x more tokens on 450M; 7.3x on 1.6B).
  Absolute percentages are inflated by the profiler's per-phase sync
  overhead — use as directional.
- WICK_PROFILE=noattn cross-check invalid for batched prefill: the env
  var only gates encode_attention / encode_attention_q_offset, not the
  batched-prefill attention kernel. Documented as follow-up in the
  report.
- Recommendation: tiled flash-attention Metal rewrite, preceded by a
  small instrumentation PR (noattn check in encode_attention_prefill +
  GPU timestamp sampling in forward_prefill_inner) for clean attribution.

Engine-side helpers for profiling:
- wick/src/model/metal_lfm2.rs: forward_prefill_profiled wrapped in a
  chunking loop matching forward_prefill_inner's pattern, so callers
  with n > MAX_PREFILL_TOKENS (512) don't overflow prefill_batch_buf.
  Upfront bounds assertion matches forward_prefill_inner to keep
  failures atomic.

Tests + tooling:
- wick/tests/bench_perf.rs: 5 new ignored profile-collection tests
  driving forward_prefill_profiled over the matrix. Shared
  aggregate_prefill_phases() helper used by both the new tests and
  the pre-existing test_prefill_phase_profile / test_450m_prefill_phase_profile
  (dedups repeated aggregation logic and fixes the empty-category row
  for phases without underscores, e.g. the whole-model "out" phase).
- scripts/profile_longctx.sh: one-shot driver reproducing every data
  point in the report (prefill matrix, decode at ctx=128/2048/4096,
  noattn wall-time cross-check).
- benchmarks/profile_longctx_raw/: 23 captured logs.
- devlog/000017-perf-profile-longctx.md: decisions, issues, discoveries.

* chore: gitignore .claude/ and remove tracked local state

.claude/ holds Claude Code local state (memory, settings, scheduled
tasks, skills) that shouldn't live in the repo. Two files were
accidentally committed on an earlier branch; remove them and add
.claude/ to .gitignore going forward.

+++ wick/src/backend/shaders/attention_gqa.metal
@@ -24,10 +24,13 @@ struct Params {
     uint _pad1;
 };
 
+// K/V caches are stored as f16 (see `encode_cast_f32_to_f16_offsets` in
+// metal_lfm2.rs and the `// f16 bytes` comment). Binding as `float*`
+// would reinterpret two adjacent halves as one f32 and produce garbage.
 kernel void attention_gqa(
     const device float* q [[buffer(0)]],
-    const device float* k_cache [[buffer(1)]],
-    const device float* v_cache [[buffer(2)]],
+    const device half*  k_cache [[buffer(1)]],
+    const device half*  v_cache [[buffer(2)]],
     device float* out [[buffer(3)]],
     constant Params& params [[buffer(4)]],
     uint tid [[thread_position_in_threadgroup]],
@@ -70,7 +73,7 @@ kernel void attention_gqa(
         float dots[MAX_GROUP];
         for (uint g = 0; g < group_size; g++) dots[g] = 0.0f;
         for (uint d = 0u; d < head_dim; d++) {
-            float k_val = k_cache[k_base + d];  // loaded once
+            float k_val = float(k_cache[k_base + d]);  // loaded once
             for (uint g = 0; g < group_size; g++) {
                 dots[g] += q_shared[g * head_dim + d] * k_val;
             }
@@ -137,7 +140,7 @@ kernel void attention_gqa(
             float val = 0.0f;
             uint s_off = g * seq_len;
             for (uint tt = s_group; tt < seq_len; tt += par) {
-                val += scores[s_off + tt] * v_cache[tt * kv_dim + kv_h_offset + d];
+                val += scores[s_off + tt] * float(v_cache[tt * kv_dim + kv_h_offset + d]);
             }
             for (uint offset = par >> 1; offset > 0u; offset >>= 1) {
                 val += simd_shuffle_down(val, offset);
